requirements.txt
# Core ML and Deep Learning
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0

# Audio Processing
librosa>=0.8.0
soundfile>=0.10.0

# Computer Vision and Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Signal Processing and Gait Analysis
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Federated Learning
flwr>=1.0.0

# Privacy and Security
cryptography>=3.4.0

# Data Handling and Storage
h5py>=3.1.0
tqdm>=4.62.0
joblib>=1.0.0

# Network and Communication
requests>=2.25.0
flask>=2.0.0

# Configuration and Logging
PyYAML>=5.4.1
python-dotenv>=0.19.0

# Testing and Quality
pytest>=6.0.0
pytest-cov>=2.12.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.0.0

# System and Performance
psutil>=5.8.0

# Additional Utilities
click>=8.0.0
pydantic>=1.8.0

# Edge Computing Simulation
networkx>=2.6.0

# Time Series Analysis
statsmodels>=0.12.0

# Model Interpretability
shap>=0.40.0

# Optimization
optuna>=2.10.0
